requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
filelock==3.12.0
greenlet==2.0.2
idna==3.4
lxml==4.9.2
numpy==1.24.3
pandas==2.0.1
platformdirs==3.4.0
PyMySQL==1.0.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
six==1.16.0
snscrape==0.6.2.20230320
soupsieve==2.4.1
SQLAlchemy==2.0.11
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.22.0
